requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
requests==2.32.3
firebase-admin==6.6.0
pytz==2024.1
